+++ src/input/Pointer.js
@@ -270,7 +270,7 @@ var Pointer = new Class({
      * @since 3.0.0
      *
      * @param {Phaser.Cameras.Scene2D.Camera} camera - [description]
-     * @param {(Phaser.Math.Vector2|object)} output - [description]
+     * @param {(Phaser.Math.Vector2|object)} [output] - [description]
      *
      * @return {(Phaser.Math.Vector2|object)} [description]
      */